modelcontextprotocol/pyproject.toml
[project]
name = "videodb-mcp-server"
version = "0.1.0"
description = "VideoDB MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "python-socketio>=5.12.1",
    "videodb>=0.2.11",
    "websocket-client>=1.8.0",
]
